adding text view and lienear layout to main
