+++ android/app/src/main/kotlin/ca/etsmtl/applets/etsmobile/presentation/dashboard/card/todayschedule/TodayScheduleCardFragment.kt
@@ -72,6 +72,7 @@ class TodayScheduleCardFragment : DaggerFragment() {
 
         todayScheduleCardViewModel.errorMessage.observe(this, EventObserver {
             it?.let {
+                if(!it.equals(context?.getString(R.string.error_no_internet_connection)))
                 Toast.makeText(context, it, Toast.LENGTH_LONG).show() }
         })
