+++ android/app/src/main/res/layout/activity_main.xml
@@ -30,7 +30,13 @@
                 android:layout_width="match_parent"
                 android:layout_height="?android:attr/actionBarSize"
                 app:titleTextColor="@color/colorToolbarAccent"
-                tools:title="Title" />
+                tools:title="Title" >
+                <LinearLayout
+                    android:id="@+id/networkMessageContainer"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:orientation="horizontal" />
+            </androidx.appcompat.widget.Toolbar>
         </com.google.android.material.appbar.CollapsingToolbarLayout>
 
         <com.google.android.material.tabs.TabLayout
